refactor: shared->shared/um, ku->shared/ku
