+++ src/dll/library.h
@@ -25,7 +25,7 @@
 #define WINFSP_DLL_INTERNAL
 #include <winfsp/winfsp.h>
 #include <winfsp/launch.h>
-#include <shared/minimal.h>
+#include <shared/um/minimal.h>
 #include <strsafe.h>
 
 #define FSP_CFG_REJECT_EARLY_IRP
